Update ChangeLog.
Compiling a changelog for 2.9.0 -> 2.9.71 is not feasible due to the immense
amount of changes and the fact that they happened multiple years ago on a
timespan of multiple years. The listing for upcoming 2.9.76, however, should
provide a somewhat reasonable representation of the changes since 2.9.71,
based on git log and coupled with relevant bug report IDs where possible.