Add a simple test for playlist layouts
This wouldn't have uncovered the bugs fixed in earlier commits as only
the LayoutManager is tested, and those were deep in the interactions between
LayoutManager and UI. Still, something is probably better than nothing.

The playlist layout management and editing code feels like something that could
benefit from being rewritten with more model-view, but that'll be later if ever.