Add workaround to enable last.fm debugging.
liblastfm checks QCoreApplication::arguments() for "--debug". That's not
easy to avoid, so add a synonym debug flag "--debug-last-fm" which has identical
functionality in Amarok, but doesn't trigger the non-functional last.fm debug
mode.

This is not pretty, but it is pretty contained and makes reasonable last.fm
bughunting possible. Should https://github.com/lastfm/liblastfm/pull/37 ever
get merged, this commit can be reverted, but that might not happen anytime soon
as the PR has been pending for 4 years and latest liblastfm release was almost
10 years ago.

BUG: 485415