Fix Wikipedia album and composer loading, add warning message option
EngineController::isStream() is broken, so just try always prettyName() for album and
composer, as the code seems to return before that anyhow if the name is empty.

The warning message was set in engine but not yet displayed in UI. Add a
Kirigami.InlineMessage for displaying it if it exists. Try to empty it when not relevant
any more, display WebEngineView only when no message set. Seemed to work ok in every
scenario I tried.