Add configuration option for showing artist in various artist cases
As demonstrated by the bug reports, this is something where one solution does not
fit all. So add a configuration option; probably - and hopefully - the users who want
to see artist in various artists compilations also want to see them when artist
differs from album artist, and vice versa. Options should be added with care, but
I think this one is definitely justified.

BUG: 276039

BUG: 284101