Avoid value and m_timeEdit->value() from getting out of sync
As m_timeEdit has a maximum of 600 set, having a higher m_filter.numValue causes
the filter and UI values to differ. Handle by forcing the 600-capped value to
m_filter if filter condition is changed. Affects collection browser search and dynamic
playlist rules, if date type is changed to older or newer than, and the default value
of 600 is not changed.