Remove non-functional Q_WS_X11 condition
According to a comment, the call was to prevent a crash on exit with Phonon-VLC
and LibPulse. The Q_WS_X11 codepath has not been triggered in Qt5 Amarok, so
apparently it was not needed any more. I also verified by running through gdb
that there's no crash at exit at the moment with Phonon-VLC + Pulseaudio combination.