Remove Q_WIN_X11 precompiler checks to enable some previous functionality
This should add some optimization to some demanding rendering paths, and allow
raising an existing Amarok window if a start of another instance is tried.
Many of the paths used have been deprecated in late KF5 versions, but it's ok.
Some don't yet seem to work, though. The activation pathways of hidden window
show some requires-attention hint, so maybe it is better to have them than not.
Additionally, only stopping rendering when different virtual desktop is shown
is pretty limited, and more optimizations could be done.