Try harder to avoid volume spikes when ReplayGain changes from low to high
Gain was adjusted in slotNewTrackPlaying, which will be called only after
the new track has started playing, causing a volume spike of a second or so.
Try to avoid this by only doing gapless playback when the next track is
from the same album, and by setting gain a bit earlier in playback preparation
if possible.

Shouldn't affect any cases where ReplayGain is not available (= phonon-vlc)
or used.

BUG: 299461