Some theoretical mp3tunes fixes
Fix the regexes so it compiles again. Add fixes to reported bugs, including a patch
posted by Zhouyang on bugs.kde.org. Fix some issues reported by code quality scans.

The compilation works now, but no effort was made even to try to test the functionality,
which probably doesn't. The service as is would need work and everything is turned off
in CMakeLists. Hard to predict if someone will ever update the mp3tunes to usable state,
but I guess there's no point in NOT applying the potential fixes (as the service still
exists, so maybe not everything's not broken enough to justify removing) so the existing
issues can be happily cleaned up at least.

BUG: 419132

BUG: 391146