Update last.fm login mechanism to use non-deprecated token method
The mobile session seems to be somewhat deprecated, and at least I'm not totally
confident that it is actually possible to log in new sessions with it any more in
Amarok.

Token-based method doesn't need password, so no new data is saved to kwallet in any
case. Tries to use existing session key from config file and username from config
file or kwallet. If username is saved to kwallet, tries to migrate it from there.
If this succeeds, everything should be transparent to user. If not, an error window
is shown and the account can be reconnected in plugin settings.

Thus also avoids any KWallet related bugs

BUG: 414826